Implement dynamic chat container styling
This commit introduced a few new changes to the way chat containers and their sections behave. CSS styles were updated to include a transition, and the chat view container width was changed from a fixed value to a dynamic one, allowing it to adapt to the content. New keyboard shortcuts (ctrl/meta + j / u) were added to the chat.js file to toggle the visibility of the user and channel containers. This addition will provide a better user experience by allowing users to have more flexibility in browsing through the chat.

In addition, the 'chat_home' component was added to offer a more user-friendly interface for users who might not have selected a channel or who are in a community without channels. The component provides a welcome message and prompts for browsing communities or creating a new one. It also hides any "example message" initially.
Lastly, some element id names have been changed for more consistency and a clearer representation of the exact function/usage of the container.

Signed-off-by: BeauTheBeau <beauthebeau@protonmail.me>